refactor(utils): move helper.ts into utils
separates from lib/ which are for 3rd party deps setup
